Se modifican algunos DTOs y se crea en su totalidad los Resources

+++ RegistroAcademico/src/main/java/registro/registroacademico/logic/CursoGrupoLogic.java
@@ -56,7 +56,7 @@ public class CursoGrupoLogic {
      * @return 
      */
     
-    public CursoGrupoEntity updateCursoGrupo(int id, CursoGrupoEntity cursoGrupoEntity){
+    public CursoGrupoEntity updateCursoGrupo(Long id, CursoGrupoEntity cursoGrupoEntity){
      CursoGrupoEntity cursoGrupo = cursoGrupoPersistence.update(cursoGrupoEntity); 
      return cursoGrupo;
     }
